Fold two copied Core test suites back into one each
PatternRegistryTests was a copy of RegistryTests: same three tests, same
names, same bodies, differing only in a stray second clear() after the
assertions that already proved the registry empty. Deleted.

PatternDetectorTests overlapped DetectorTests on four of its six tests.
Its two unique ones move across. Its copy of testCrossFilePatternDetection
was the stronger of the pair and is the one kept: both of its fixture views
declare the same @State variable, which is the shape
.relatedDuplicateStateVariable looks for, where DetectorTests used two
stateless views and so could not tell "quiet here" from "never fires".

3,479 tests in 420 suites pass — seven fewer tests, all of them second
copies of a test that still runs.
