--- Tests/CoreTests/SwiftSyntaxPattern/PatternRegistryTests.swift
@@ -1,134 +1,0 @@
-@testable import Core
-///
-/// NOTE ON TEST ISOLATION
-///
-/// These tests use isolated registry instances to ensure complete test isolation
-/// and avoid interference between tests.
-///
-import Foundation
-import SwiftParser
-@testable import SwiftProjectLintRules
-import SwiftSyntax
-import Testing
-
-struct PatternRegistryTests {
-
-    // MARK: - Test Helper Methods
-
-    /// Creates isolated instances for tests that need complete isolation
-    @MainActor static func createIsolatedInstances() -> IsolatedTestInstances {
-        TestRegistryManager.createIsolatedInstances()
-    }
-
-    // MARK: - Registry Tests (Need Isolation)
-
-    @Test
-    @MainActor
-    static func patternVisitorRegistryRegistration() throws {
-        let instances = createIsolatedInstances()
-        let testVisitorRegistry = instances.visitorRegistry
-
-        // Given
-        let pattern = SyntaxPattern(
-            name: .fatView,
-            visitor: SwiftUIManagementVisitor.self,
-            severity: .warning,
-            category: .stateManagement,
-            messageTemplate: "Test message",
-            suggestion: "Test suggestion",
-            description: "Test description"
-        )
-
-        // When
-        testVisitorRegistry.register(pattern: pattern)
-
-        // Then
-        let patterns = testVisitorRegistry.getAllPatterns()
-        let registeredPattern = try #require(patterns.first)
-        #expect(registeredPattern.name == .fatView)
-
-        let visitors = testVisitorRegistry.getVisitors(for: .stateManagement)
-        let registeredVisitor = try #require(visitors.first)
-        #expect(registeredVisitor is SwiftUIManagementVisitor.Type)
-
-        testVisitorRegistry.clear()
-    }
-
-    @Test
-    @MainActor
-    static func patternVisitorRegistryMultiplePatterns() {
-        let instances = createIsolatedInstances()
-        let testVisitorRegistry = instances.visitorRegistry
-
-        // Given
-        let patterns = [
-            SyntaxPattern(
-                name: .fatView,
-                visitor: SwiftUIManagementVisitor.self,
-                severity: .warning,
-                category: .stateManagement,
-                messageTemplate: "Message 1",
-                suggestion: "Suggestion 1",
-                description: "Description 1"
-            ),
-            SyntaxPattern(
-                name: .uninitializedStateVariable,
-                visitor: SwiftUIManagementVisitor.self,
-                severity: .error,
-                category: .stateManagement,
-                messageTemplate: "Message 2",
-                suggestion: "Suggestion 2",
-                description: "Description 2"
-            )
-        ]
-
-        // When
-        testVisitorRegistry.register(patterns: patterns)
-
-        // Then - Check that our specific patterns are registered
-        let allPatterns = testVisitorRegistry.getAllPatterns()
-        let ourPatterns = allPatterns.filter { pattern in
-            patterns.contains { $0.name == pattern.name }
-        }
-        #expect(ourPatterns.count == 2)
-
-        let stateManagementPatterns = testVisitorRegistry.getPatterns(for: .stateManagement)
-        let ourStatePatterns = stateManagementPatterns.filter { pattern in
-            patterns.contains { $0.name == pattern.name }
-        }
-        #expect(ourStatePatterns.count == 2)
-
-        let visitors = testVisitorRegistry.getVisitors(for: .stateManagement)
-        #expect(visitors.count >= 2) // At least our 2 visitors
-
-        testVisitorRegistry.clear()
-    }
-
-    @Test
-    @MainActor
-    static func patternVisitorRegistryClear() {
-        let instances = createIsolatedInstances()
-        let testVisitorRegistry = instances.visitorRegistry
-
-        // Given
-        let pattern = SyntaxPattern(
-            name: .fatView,
-            visitor: SwiftUIManagementVisitor.self,
-            severity: .warning,
-            category: .stateManagement,
-            messageTemplate: "Test message",
-            suggestion: "Test suggestion",
-            description: "Test description"
-        )
-        testVisitorRegistry.register(pattern: pattern)
-
-        // When
-        testVisitorRegistry.clear()
-
-        // Then
-        #expect(testVisitorRegistry.getAllPatterns().isEmpty)
-        #expect(testVisitorRegistry.getVisitors(for: .stateManagement).isEmpty)
-
-        testVisitorRegistry.clear()
-    }
-}
